requirements.txt
anthropic
python-telegram-bot
apscheduler
requests
